test: Add tests for recovery
Also adds a new test for recovery from a swap that has been claimed by
csv in the time the node was shut down.
